+++ CHANGES.md
@@ -1,6 +1,15 @@
 videojs-record changelog
 ========================
 
+2.2.0 - unreleased
+------------------
+
+- Package library using webpack (#226)
+- Add unit tests using karma/webpack/jasmine (#226)
+- Bump required version for videojs-wavesurfer to 2.3.0 for wavesurfer
+  destroy fixes
+
+
 2.1.3 - 2018/04/08
 ------------------